Update: 画面のレイアウトを調整
